[AC-1174] Reorganize authorization handlers in Core project
